Updated with images and column organization
The extra code is meant to separate the pictures and links into columns in their respective rows to space them out.  It doesn't look quite that way on the site yet, but I plan to make that look more complete in due time, I just wanted to put the code in for now so I can finish it up later.